make grid
In this grid, each cell was given an id containing its [the cell's] respective coordinate. Each cell's id is of the form (row, col); row represents the row in which it is located, whereas col represents the column in which it is located.

Moreover, each cell was given the class "unvisited_node".