fix astar algorithm
Before this change, my astar algorithm considered only the respective manhattan distances of each node to determine which given node was going to be spliced next from the queue; therefore, I needed to change the algorithm such that it considered the heuristic value (the value of a node's "f" attribute) to decide which node to splice from the queue.

Additionally, this commit enables different paths to be visualize on the same configuration of walls. Previously, I had to generate new walls to visualize an algorithm.